requirements/dev.txt
factory_boy==3.3.3
Faker==37.1.0
isort==5.13.2
pylint==3.3.3
pycodestyle==2.12.1
pylint_django==2.6.1
pytest==8.3.5
pytest-django==4.10.0
pytest-cov==6.0.0
pytz==2025.2
tox==4.24.1